fix rag_chain object to be compatiple with langchain, organize code file
