+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+	"python-envs.defaultEnvManager": "ms-python.python:conda",
+		"python-envs.defaultPackageManager": "ms-python.python:conda",
+			"python-envs.pythonProjects": []
+}
