+++ assets/feature/.gitignore
@@ -1,0 +1,1 @@
+token.txt
