+++ assets/feature/rag_chatbot.py
@@ -1,3 +1,4 @@
+from langchain_huggingface import HuggingFacePipeline
 import streamlit as st
 import os
 import requests
@@ -18,6 +19,12 @@ import time
 import tempfile
 import urllib.parse
 import zipfile
+from langchain_core.runnables import RunnableLambda
+import torch
+from transformers import AutoModelForCausalLM, AutoTokenizer
+
+
+
 
 st.set_page_config(
     page_title="Trợ Lý AI Tiếng Việt",
@@ -162,10 +169,6 @@ if 'chat_history' not in st.session_state:
     st.session_state.chat_history = []
 if 'rag_chain' not in st.session_state:
     st.session_state.rag_chain = None
-if 'models_loaded' not in st.session_state:
-    st.session_state.models_loaded = False
-if 'embeddings' not in st.session_state:
-    st.session_state.embeddings = None
 if 'documents_loaded' not in st.session_state:
     st.session_state.documents_loaded = False
 if 'pdf_source' not in st.session_state:
@@ -179,257 +182,53 @@ if 'processing_query' not in st.session_state:
 if 'query_input' not in st.session_state:
     st.session_state.query_input = ""
 
-@st.cache_resource
-def load_embeddings():
-    """Tải mô hình embedding tiếng Việt"""
-    return HuggingFaceEmbeddings(model_name="bkai-foundation-models/vietnamese-bi-encoder")
+# Check if models downloaded or not
+if 'models_loaded' not in st.session_state:
+    st.session_state.models_loaded = False
 
-def simple_text_generation(question, context):
-    """Fallback function sử dụng rule-based approach với debugging"""
-    try:
-        # Kiểm tra context
-        if not context or len(context.strip()) < 10:
-            return "Nội dung tài liệu không đủ để trả lời câu hỏi."
-        
-        # Tìm kiếm từ khóa trong context
-        question_lower = question.lower()
-        context_lower = context.lower()
-        
-        # Chia context thành câu
-        sentences = []
-        # Thử nhiều cách chia câu
-        for delimiter in ['. ', '.\n', '! ', '?\n', '? ']:
-            if delimiter in context:
-                sentences.extend(context.split(delimiter))
-        
-        # Nếu không có câu nào, chia theo đoạn
-        if not sentences:
-            sentences = context.split('\n')
-        
-        # Lọc câu rỗng
-        sentences = [s.strip() for s in sentences if len(s.strip()) > 20]
-        
-        if not sentences:
-            return "Không thể phân tích cấu trúc tài liệu. Có thể tài liệu bị lỗi format."
-        
-        # Từ dừng tiếng Việt mở rộng
-        vietnamese_stopwords = {
-            'là', 'của', 'và', 'với', 'cho', 'từ', 'về', 'theo', 'trong', 'nào', 'gì', 'sao', 
-            'thế', 'như', 'có', 'không', 'được', 'này', 'đó', 'những', 'các', 'một', 'hai', 
-            'ba', 'bốn', 'năm', 'sáu', 'bảy', 'tám', 'chín', 'mười', 'the', 'and', 'or', 
-            'but', 'in', 'on', 'at', 'to', 'for', 'of', 'with', 'by', 'what', 'how', 'when', 
-            'where', 'why', 'who', 'which', 'that', 'this', 'these', 'those', 'will', 'would',
-            'could', 'should', 'may', 'might', 'can', 'must', 'shall'
-        }
-        
-        # Trích xuất từ khóa từ câu hỏi
-        question_words = []
-        for word in question_lower.split():
-            clean_word = word.strip('.,!?()[]{}":;').lower()
-            if len(clean_word) > 2 and clean_word not in vietnamese_stopwords:
-                question_words.append(clean_word)
-        
-        if not question_words:
-            return "Không thể xác định từ khóa từ câu hỏi. Vui lòng đặt câu hỏi cụ thể hơn."
-        
-        # Tìm câu liên quan
-        relevant_sentences = []
-        for sentence in sentences:
-            sentence_lower = sentence.lower()
-            score = 0
-            
-            # Tính điểm dựa trên số từ khóa xuất hiện
-            for word in question_words:
-                if word in sentence_lower:
-                    # Từ xuất hiện chính xác
-                    score += 2
-                    # Bonus nếu từ xuất hiện nhiều lần
-                    score += sentence_lower.count(word) - 1
-            
-            # Bonus cho câu chứa nhiều từ khóa
-            if score > 0:
-                word_coverage = sum(1 for word in question_words if word in sentence_lower)
-                coverage_bonus = (word_coverage / len(question_words)) * 2
-                score += coverage_bonus
-                
-                relevant_sentences.append((sentence.strip(), score))
-        
-        if relevant_sentences:
-            # Sắp xếp theo điểm relevance và lấy top 5
-            relevant_sentences.sort(key=lambda x: x[1], reverse=True)
-            top_sentences = [sent[0] for sent in relevant_sentences[:5]]
-            
-            answer = "Dựa trên tài liệu, tôi tìm thấy thông tin sau:\n\n"
-            for i, sentence in enumerate(top_sentences, 1):
-                if sentence.strip():
-                    # Làm sạch câu
-                    clean_sentence = sentence.strip()
-                    if not clean_sentence.endswith(('.', '!', '?')):
-                        clean_sentence += '.'
-                    answer += f"{i}. {clean_sentence}\n\n"
-            
-            return answer.strip()
-        else:
-            # Fallback: trả về một phần ngẫu nhiên của text
-            preview_text = context[:1000] + "..." if len(context) > 1000 else context
-            return f"Tôi không thể tìm thấy thông tin cụ thể liên quan đến câu hỏi '{question}' trong tài liệu. Tuy nhiên, đây là một phần nội dung tài liệu:\n\n{preview_text}\n\nVui lòng thử diễn đạt lại câu hỏi hoặc hỏi về các chủ đề khác được đề cập trong tài liệu."
-    
-    except Exception as e:
-        return f"Xin lỗi, tôi gặp lỗi khi tìm kiếm: {str(e)}. Vui lòng thử lại."
+# save downloaded embeding model
+if 'embeddings' not in st.session_state:
+    st.session_state.embeddings = None
 
-def extract_text_from_uploaded_file(file):
-    """Trích xuất văn bản từ file được tải lên dựa trên loại file"""
-    file_extension = file.name.split('.')[-1].lower()
-    
-    try:
-        # Tạo file tạm thời
-        with tempfile.NamedTemporaryFile(delete=False, suffix=f'.{file_extension}') as tmp_file:
-            tmp_file.write(file.getbuffer())
-            tmp_path = tmp_file.name
-        
-        documents = []
-        
-        if file_extension == 'pdf':
-            loader = PyPDFLoader(tmp_path)
-            documents = loader.load()
-        elif file_extension == 'docx':
-            loader = Docx2txtLoader(tmp_path)
-            documents = loader.load()
-        elif file_extension in ['xlsx', 'xls']:
-            loader = UnstructuredExcelLoader(tmp_path)
-            documents = loader.load()
-        else:
-            st.warning(f"Định dạng file không được hỗ trợ: {file_extension}")
-            return []
-        
-        # Dọn dẹp file tạm thời
-        os.unlink(tmp_path)
-        return documents
-        
-    except Exception as e:
-        st.error(f"Lỗi khi xử lý {file.name}: {str(e)}")
-        if 'tmp_path' in locals() and os.path.exists(tmp_path):
-            os.unlink(tmp_path)
-        return []
+# Save downloaded LLM
+if 'llm' not in st.session_state:
+    st.session_state.llm = None
 
-def process_zip_file(zip_file):
-    """Xử lý file zip được tải lên chứa tài liệu"""
-    try:
-        all_documents = []
-        loaded_files = []
-        
-        # Tạo thư mục tạm thời để giải nén
-        temp_dir = tempfile.mkdtemp()
-        
-        # Lưu file zip được tải lên
-        zip_path = os.path.join(temp_dir, zip_file.name)
-        with open(zip_path, 'wb') as f:
-            f.write(zip_file.getbuffer())
-        
-        # Giải nén file zip
-        with zipfile.ZipFile(zip_path, 'r') as zip_ref:
-            zip_ref.extractall(temp_dir)
-        
-        # Tìm tất cả file tài liệu trong thư mục đã giải nén
-        for root, dirs, files in os.walk(temp_dir):
-            for file in files:
-                if file.lower().endswith(('.pdf', '.docx', '.xlsx', '.xls')):
-                    file_path = os.path.join(root, file)
-                    try:
-                        if file.lower().endswith('.pdf'):
-                            loader = PyPDFLoader(file_path)
-                        elif file.lower().endswith('.docx'):
-                            loader = Docx2txtLoader(file_path)
-                        elif file.lower().endswith(('.xlsx', '.xls')):
-                            loader = UnstructuredExcelLoader(file_path)
-                        
-                        documents = loader.load()
-                        all_documents.extend(documents)
-                        loaded_files.append(file)
-                        st.success(f"✅ Đã xử lý từ zip: {file}")
-                        
-                    except Exception as e:
-                        st.error(f"❌ Lỗi khi xử lý {file} từ zip: {str(e)}")
-        
-        # Dọn dẹp
-        shutil.rmtree(temp_dir)
-        return all_documents, loaded_files
-        
-    except Exception as e:
-        st.error(f"Lỗi khi xử lý file zip: {str(e)}")
-        return [], []
+# Import file processing function from process_file.py
+from process_file import *
+from load_llm import *
 
-def get_github_pdf_files(repo_url):
-    """Lấy danh sách file PDF từ GitHub repository"""
-    try:
-        if "github.com" in repo_url and "/tree/" in repo_url:
-            parts = repo_url.replace("https://github.com/", "").split("/tree/")
-            repo_path = parts[0]
-            branch_and_path = parts[1].split("/", 1)
-            branch = branch_and_path[0]
-            folder_path = branch_and_path[1] if len(branch_and_path) > 1 else ""
+def format_docs(docs):
+    if not docs:
+        return "Không tìm thấy tài liệu liên quan."
+    return "\n\n".join(doc.page_content for doc in docs)
 
-            api_url = f"https://api.github.com/repos/{repo_path}/contents/{folder_path}?ref={branch}"
-        else:
-            st.error("Định dạng URL GitHub không hợp lệ")
-            return []
 
-        response = requests.get(api_url)
-        if response.status_code == 200:
-            files = response.json()
-            pdf_files = []
-            for file in files:
-                if file['name'].endswith('.pdf') and file['type'] == 'file':
-                    pdf_files.append({
-                        'name': file['name'],
-                        'download_url': file['download_url']
-                    })
-            return pdf_files
-        else:
-            st.error(f"Không thể truy cập GitHub repository: {response.status_code}")
-            return []
-    except Exception as e:
-        st.error(f"Lỗi khi truy cập GitHub repository: {str(e)}")
-        return []
-
-def download_pdf_from_url(url, filename, temp_dir):
-    """Tải file PDF từ URL"""
-    try:
-        response = requests.get(url)
-        if response.status_code == 200:
-            file_path = os.path.join(temp_dir, filename)
-            with open(file_path, 'wb') as f:
-                f.write(response.content)
-            return file_path
-        return None
-    except Exception as e:
-        st.error(f"Lỗi khi tải {filename}: {str(e)}")
-        return None
 
 def create_rag_chain(all_documents):
     """Tạo chuỗi RAG từ tài liệu"""
     if not all_documents:
         st.error("Không có tài liệu nào để xử lý")
         return None, 0
-    
+
     try:
         st.info(f"🔄 Đang xử lý {len(all_documents)} tài liệu...")
-        
+
         # Kiểm tra nội dung tài liệu
         total_text = ""
         for doc in all_documents:
             if hasattr(doc, 'page_content'):
                 total_text += doc.page_content + "\n"
-        
+
         if len(total_text.strip()) < 50:
             st.error("Nội dung tài liệu quá ngắn hoặc không thể đọc được")
             return None, 0
-        
+
         st.success(f"✅ Đã đọc {len(total_text):,} ký tự từ tài liệu")
-        
+
         # Lưu toàn bộ text vào session state để fallback
         st.session_state.documents_text = total_text
-        
+
         # Sử dụng text splitter mạnh mẽ hơn nếu SemanticChunker thất bại
         try:
             if st.session_state.embeddings:
@@ -456,19 +255,19 @@ def create_rag_chain(all_documents):
             )
             docs = text_splitter.split_documents(all_documents)
             st.info(f"✅ Sử dụng RecursiveCharacterTextSplitter: {len(docs)} chunks")
-        
+
         if not docs:
             st.error("Không có đoạn tài liệu nào được tạo")
             # Tạo simple RAG chain với toàn bộ text
             def simple_rag_chain_text(question):
-                return simple_text_generation(question, total_text)
+                return simple_text_retrieval(question, total_text)
             return simple_rag_chain_text, 1
-        
+
         # Triển khai FAISS với xử lý lỗi (chỉ khi có embeddings)
         if st.session_state.embeddings:
             try:
                 vector_db = FAISS.from_documents(documents=docs, embedding=st.session_state.embeddings)
-                retriever = vector_db.as_retriever(search_kwargs={"k": min(4, len(docs))})
+                retriever = vector_db.as_retriever(top_k=5)
                 st.success(f"✅ Đã tạo FAISS vector database với {len(docs)} chunks")
             except Exception as e:
                 st.error(f"Lỗi khi tạo FAISS vector database: {str(e)}")
@@ -476,67 +275,103 @@ def create_rag_chain(all_documents):
                 # Fallback to simple text search
                 def simple_rag_chain_docs(question):
                     combined_text = "\n\n".join([doc.page_content for doc in docs])
-                    return simple_text_generation(question, combined_text)
+                    return simple_text_retrieval(question, combined_text)
                 return simple_rag_chain_docs, len(docs)
         else:
             st.info("🔍 Không có embeddings, sử dụng tìm kiếm text đơn giản")
             def simple_rag_chain_docs(question):
                 combined_text = "\n\n".join([doc.page_content for doc in docs])
-                return simple_text_generation(question, combined_text)
+                return simple_text_retrieval(question, combined_text)
             return simple_rag_chain_docs, len(docs)
 
         # Sử dụng tìm kiếm từ khóa thông minh với hub prompt
         st.info("🔍 Sử dụng tìm kiếm từ khóa thông minh với RAG prompt")
-        
+
+        #? Code dư thừa: prompt trong link rlm/rag-prompt với prompt cục bộ giống nhau
         # Tải prompt từ hub
-        try:
-            prompt = hub.pull("rlm/rag-prompt")
-            st.success("✅ Đã tải prompt template từ hub")
-        except Exception as e:
-            st.warning(f"Không thể tải prompt từ hub: {str(e)}")
-            st.info("🔄 Sử dụng prompt template cục bộ...")
-            prompt_template = """Sử dụng những đoạn ngữ cảnh sau để trả lời câu hỏi ở cuối.
-            Nếu bạn không biết câu trả lời, chỉ cần nói rằng bạn không biết, đừng cố bịa ra câu trả lời.
-            Trả lời bằng tiếng Việt.
+        # try:
+        #     prompt = hub.pull("rlm/rag-prompt")
+        #     st.success("✅ Đã tải prompt template từ hub")
+        # except Exception as e:
+        # st.warning(f"Không thể tải prompt từ hub: {str(e)}")
+        st.info("🔄 Sử dụng prompt template cục bộ...")
 
-            Ngữ cảnh: {context}
+        prompt = """Sử dụng những đoạn ngữ cảnh sau để trả lời câu hỏi ở cuối.
+        Nếu bạn không biết câu trả lời, chỉ cần nói rằng bạn không biết, đừng cố bịa ra câu trả lời.
+        Trả lời bằng tiếng Việt.
 
-            Câu hỏi: {question}
+        Ngữ cảnh: {context}
 
-            Trả lời:"""
-            prompt = PromptTemplate(
-                template=prompt_template,
-                input_variables=["context", "question"]
-            )
+        Câu hỏi: {question}
 
-        def format_docs(docs):
-            if not docs:
-                return "Không tìm thấy tài liệu liên quan."
-            return "\n\n".join(doc.page_content for doc in docs)
-        
+        Trả lời:
+        """
+
+        prompt_template = PromptTemplate(
+            template=prompt,
+            input_variables=["context", "question"]
+        )
+
+
+        rag_chain = (
+            {
+                "context": retriever | format_docs,
+                "question": RunnablePassthrough()
+            }
+            | prompt_template
+            | st.session_state.llm
+            | StrOutputParser()
+        )
+        st.write(f"___[DEBUG]__\n")
+
         # Tạo simple RAG chain sử dụng keyword search với prompt
-        def smart_rag_chain_with_prompt(question):
-            try:
-                # Tìm tài liệu liên quan bằng retriever
-                relevant_docs = retriever.get_relevant_documents(question)
-                context = format_docs(relevant_docs)
-                
-                # Sử dụng simple text generation để trả lời
-                return simple_text_generation(question, context)
-                
-            except Exception as e:
-                st.warning(f"Lỗi retriever: {str(e)}, sử dụng toàn bộ text")
-                return simple_text_generation(question, total_text)
-        
-        return smart_rag_chain_with_prompt, len(docs)
-        
+        # def smart_rag_chain_with_prompt(question):
+        #     try:
+        #         # Tìm tài liệu liên quan bằng retriever
+        #         relevant_docs = retriever.get_relevant_documents(question)
+        #         context = format_docs(relevant_docs)
+
+        #         # Sử dụng simple text generation để trả lời
+        #         context = simple_text_retrieval(question, context)
+
+        #         rag_chain = (
+        #             {
+        #                 "context": RunnableLambda(lambda _: context),
+        #                 "question": RunnablePassthrough()
+        #             }
+        #             | prompt
+        #             | st.session_state.llm
+        #             | StrOutputParser()
+        #         )
+
+        #         return rag_chain
+
+        #     except Exception as e:
+        #         st.warning(f"Lỗi retriever: {str(e)}, sử dụng toàn bộ text")
+        #         context = simple_text_retrieval(question, total_text)
+
+        #         rag_chain = (
+        #             {
+        #                 "context": RunnableLambda(lambda _: context),
+        #                 "question": RunnablePassthrough()
+        #             }
+        #             | prompt
+        #             | st.session_state.llm
+        #             | StrOutputParser()
+        #         )
+        #         st.write(f"[DEBUG] Lỗi retriever fallback: {str(e)}")
+
+        #         return rag_chain, len(docs)
+
+        return rag_chain, len(docs) # basically return rag_chain, len(docs)
+
     except Exception as e:
         st.error(f"Lỗi nghiêm trọng khi tạo chuỗi RAG: {str(e)}")
         st.info("🔄 Tạo fallback RAG chain...")
         # Ultimate fallback
         def emergency_rag_chain(question):
             if hasattr(st.session_state, 'documents_text') and st.session_state.documents_text:
-                return simple_text_generation(question, st.session_state.documents_text)
+                return simple_text_retrieval(question, st.session_state.documents_text)
             else:
                 return "Xin lỗi, không thể truy cập nội dung tài liệu. Vui lòng tải lại tài liệu."
         return emergency_rag_chain, 1
@@ -649,16 +484,17 @@ def display_thinking_indicator():
     </div>
     """, unsafe_allow_html=True)
 
+#? rag_chain.invoke typeof function
 def process_user_query(question):
     """Xử lý câu hỏi của người dùng"""
     try:
         if not st.session_state.rag_chain:
             return "Xin lỗi, chưa có tài liệu nào được tải. Vui lòng tải lên hoặc nạp tài liệu trước."
-        
+
         # Kiểm tra câu hỏi
         if not question or len(question.strip()) < 2:
             return "Vui lòng đặt câu hỏi cụ thể hơn."
-        
+
         # Gọi chuỗi RAG với xử lý lỗi chi tiết
         try:
             if callable(st.session_state.rag_chain):
@@ -667,15 +503,15 @@ def process_user_query(question):
             else:
                 # LangChain RAG chain (không có vì đã bỏ LLM)
                 output = st.session_state.rag_chain(question)
-            
+
         except Exception as chain_error:
             st.error(f"Lỗi khi gọi RAG chain: {str(chain_error)}")
             # Ultimate fallback: sử dụng documents_text nếu có
             if hasattr(st.session_state, 'documents_text') and st.session_state.documents_text:
-                return simple_text_generation(question, st.session_state.documents_text)
+                return simple_text_retrieval(question, st.session_state.documents_text)
             else:
                 return f"Xin lỗi, gặp lỗi khi xử lý câu hỏi: {str(chain_error)}. Vui lòng thử tải lại tài liệu."
-        
+
         # Xử lý các định dạng đầu ra khác nhau
         if isinstance(output, str):
             # Nếu đầu ra chứa "Answer:", trích xuất phần sau nó
@@ -696,21 +532,21 @@ def process_user_query(question):
         else:
             # Nếu đầu ra không phải là chuỗi, chuyển đổi nó
             answer = str(output).strip()
-        
+
         # Đảm bảo có câu trả lời có ý nghĩa
         if not answer or len(answer) < 5:
             return "Tôi đã tìm thấy một số thông tin trong tài liệu, nhưng không thể tạo ra câu trả lời rõ ràng. Vui lòng thử diễn đạt lại câu hỏi của bạn."
-        
+
         # Làm sạch câu trả lời
         answer = answer.replace("Human:", "").replace("Assistant:", "").strip()
-        
+
         return answer
-        
+
     except Exception as e:
         st.error(f"Lỗi không mong đợi: {str(e)}")
         # Thử fallback cuối cùng
         if hasattr(st.session_state, 'documents_text') and st.session_state.documents_text:
-            return simple_text_generation(question, st.session_state.documents_text)
+            return simple_text_retrieval(question, st.session_state.documents_text)
         return "Tôi xin lỗi, gặp lỗi không mong đợi. Vui lòng thử tải lại tài liệu hoặc đặt câu hỏi khác."
 
 def main():
@@ -751,7 +587,7 @@ def main():
 
         if pdf_source == "Tải File Lên":
             st.session_state.pdf_source = "upload_files"
-            
+
             st.markdown('<div class="upload-section">', unsafe_allow_html=True)
             st.markdown("**📎 Tải Lên Từng File**")
             uploaded_files = st.file_uploader(
@@ -761,20 +597,20 @@ def main():
                 help="Định dạng hỗ trợ: PDF, Word (.docx), Excel (.xlsx, .xls)"
             )
             st.markdown('</div>', unsafe_allow_html=True)
-            
+
             if uploaded_files:
                 st.markdown("**File Đã Chọn:**")
                 for i, file in enumerate(uploaded_files):
                     file_size = len(file.getbuffer()) / (1024 * 1024)  # Kích thước tính bằng MB
                     st.markdown(f'<span class="file-counter">{i+1}. {file.name} ({file_size:.1f} MB)</span>', unsafe_allow_html=True)
-                
+
                 if st.button("📤 Xử Lý File Đã Tải", type="primary"):
                     with st.spinner("Đang xử lý file đã tải lên..."):
                         all_documents = []
                         loaded_files = []
-                        
+
                         progress_bar = st.progress(0)
-                        
+
                         for i, file in enumerate(uploaded_files):
                             documents = extract_text_from_uploaded_file(file)
                             if documents:
@@ -782,9 +618,9 @@ def main():
                                 loaded_files.append(file.name)
                                 st.success(f"✅ Đã xử lý: {file.name}")
                             progress_bar.progress((i + 1) / len(uploaded_files))
-                        
+
                         progress_bar.empty()
-                        
+
                         if all_documents:
                             rag_chain, num_chunks = create_rag_chain(all_documents)
                             if rag_chain:
@@ -797,7 +633,7 @@ def main():
 
         elif pdf_source == "Tải Thư Mục (ZIP)":
             st.session_state.pdf_source = "upload_zip"
-            
+
             st.markdown('<div class="upload-section">', unsafe_allow_html=True)
             st.markdown("**📁 Tải Thư Mục Dưới Dạng ZIP**")
             zip_file = st.file_uploader(
@@ -806,15 +642,15 @@ def main():
                 help="Tải lên file ZIP chứa file PDF, Word, hoặc Excel"
             )
             st.markdown('</div>', unsafe_allow_html=True)
-            
+
             if zip_file:
                 file_size = len(zip_file.getbuffer()) / (1024 * 1024)  # Kích thước tính bằng MB
                 st.info(f"📦 File ZIP đã chọn: {zip_file.name} ({file_size:.1f} MB)")
-                
+
                 if st.button("📤 Xử Lý File ZIP", type="primary"):
                     with st.spinner("Đang giải nén và xử lý file ZIP..."):
                         all_documents, loaded_files = process_zip_file(zip_file)
-                        
+
                         if all_documents:
                             rag_chain, num_chunks = create_rag_chain(all_documents)
                             if rag_chain:
@@ -833,7 +669,7 @@ def main():
                 help="URL đến thư mục GitHub chứa file PDF"
             )
             st.session_state.github_repo_url = github_url
-            
+
             if st.button("📥 Tải Từ GitHub", type="primary"):
                 st.session_state.documents_loaded = False
                 st.rerun()
@@ -846,7 +682,7 @@ def main():
                 help="Đường dẫn đến thư mục cục bộ chứa file PDF"
             )
             st.session_state.local_folder_path = local_path
-            
+
             if st.button("📂 Tải Từ Thư Mục Cục Bộ", type="primary"):
                 st.session_state.documents_loaded = False
                 st.rerun()
@@ -873,7 +709,7 @@ def main():
         st.info("🚀 FAISS: Thư viện tìm kiếm tương tự nhanh")
         st.info("🔍 Keyword Search: Tìm kiếm từ khóa thông minh")
         st.info("🌐 Hub Prompt: Sử dụng rlm/rag-prompt template")
-        
+
         st.divider()
         st.subheader("🇻🇳 Mô Hình Tiếng Việt")
         st.markdown("""
@@ -890,11 +726,11 @@ def main():
             </p>
         </div>
         """, unsafe_allow_html=True)
-        
+
         # Debug section
         st.divider()
         st.subheader("🔧 Debug & Kiểm Tra")
-        
+
         if st.button("🔍 Kiểm Tra Hệ Thống"):
             st.write("**Trạng thái Hệ Thống:**")
             st.write(f"- Models loaded: {st.session_state.models_loaded}")
@@ -902,12 +738,12 @@ def main():
             st.write(f"- Documents loaded: {st.session_state.documents_loaded}")
             st.write(f"- RAG chain: {'✅' if st.session_state.rag_chain else '❌'}")
             st.write(f"- Mode: 🔍 Keyword Search (No API required)")
-            
+
             if hasattr(st.session_state, 'documents_text'):
                 st.write(f"- Documents text length: {len(st.session_state.documents_text):,} characters")
             else:
                 st.write("- Documents text: ❌ Chưa có")
-        
+
         if st.session_state.documents_loaded and st.button("📄 Xem Mẫu Nội Dung"):
             if hasattr(st.session_state, 'documents_text') and st.session_state.documents_text:
                 preview = st.session_state.documents_text[:500] + "..." if len(st.session_state.documents_text) > 500 else st.session_state.documents_text
@@ -923,9 +759,9 @@ def main():
                 st.error(f"❌ Lỗi khi tải embeddings: {str(e)}")
                 st.warning("⚠️ Sẽ hoạt động ở chế độ đơn giản mà không có embeddings")
                 st.session_state.embeddings = None
-            
+
             st.session_state.models_loaded = True
-        
+
         st.success("✅ Hệ thống đã sẵn sàng!")
         st.info("🔍 Đang hoạt động ở chế độ tìm kiếm từ khóa thông minh")
         time.sleep(1)
@@ -934,13 +770,17 @@ def main():
     # Tải tài liệu nếu chưa được tải và nguồn là github hoặc local
     if st.session_state.models_loaded and not st.session_state.documents_loaded and st.session_state.pdf_source in ["github", "local"]:
         with st.spinner("📚 Đang tải tài liệu vào kho vector FAISS..."):
+            #? rag_chain save into session state, and become available in every function e.g. process_user_query()
             if st.session_state.pdf_source == "github":
-                rag_chain, num_chunks, loaded_files = load_pdfs_from_github(st.session_state.github_repo_url)
-            else:
-                rag_chain, num_chunks, loaded_files = load_pdfs_from_folder(st.session_state.local_folder_path)
+                st.session_state.rag_chain, num_chunks, loaded_files = load_pdfs_from_github(st.session_state.github_repo_url)
+                print("\n---github---\n")
 
-            if rag_chain:
-                st.session_state.rag_chain = rag_chain
+            else:
+                st.session_state.rag_chain, num_chunks, loaded_files = load_pdfs_from_folder(st.session_state.local_folder_path)
+                print("\n---load from folder---\n")
+
+            if st.session_state.rag_chain:
+                # st.session_state.rag_chain = rag_chain #? Lỗi CHÍNH 2 đặt sai biến. phải ngược lại mới đúng.
                 st.session_state.documents_loaded = True
 
                 st.markdown(f"""
@@ -1011,7 +851,8 @@ def main():
                 st.session_state.processing_query = False
             else:
                 last_question = st.session_state.chat_history[-1]["content"]
-                answer = process_user_query(last_question)
+                # answer = process_user_query(last_question)
+                answer = st.session_state.rag_chain.invoke(last_question)
 
                 st.session_state.chat_history.append({
                     "content": answer,
