+++ public/posts/2025/week-4/chat_bot/index.html
@@ -220,33 +220,58 @@
   </tbody>
 </table>
 <hr>
-<p>📂 <em>Tài liệu đi kèm:</em><br>
+<p>📂 <em>Tài liệu đi kèm:</em>
 <details>
-  	<summary> Week_4_Streamlit_RAGChatbot.pdf </summary>
-  	<div class="pdf-container my-6 border rounded-lg overflow-hidden shadow-lg">
-  		<div class="pdf-header bg-zinc-100 px-4 py-2 flex items-center justify-between">
-  			<span class="font-semibold text-sm">📄 Week_4_Streamlit_RAGChatbot.pdf </span>
-  			<div class="flex space-x-2">
-  				<a href="/Time-Series-Team-Hub/pdf/Week_4___Streamlit_RAGChatbot.pdf" target="_blank" class="text-blue-600 hover:text-blue-800 text-sm">
-  					Open in New Tab
-  				</a>
-  				<a href="/Time-Series-Team-Hub/pdf/Week_4___Streamlit_RAGChatbot.pdf" download class="text-blue-600 hover:text-blue-800 text-sm">
-  					Download
-  				</a>
-  			</div>
-  		</div>
-  		<iframe 
-  			src="/Time-Series-Team-Hub/pdf/Week_4___Streamlit_RAGChatbot.pdf" 
-  			width="100%" 
-  			height="700px"
-  			class="border-0"
-  			title="Week_4_Streamlit_RAGChatbot.pdf">
-  			<p>Your browser does not support PDFs. <a href="/Time-Series-Team-Hub/pdf/Week_4___Streamlit_RAGChatbot.pdf">Download the PDF</a>.</p>
-  		</iframe>
-  	</div>
-  </details>
+	<summary> Week_4_Streamlit_RAGChatbot.pdf </summary>
+	<div class="pdf-container my-6 border rounded-lg overflow-hidden shadow-lg">
+		<div class="pdf-header bg-zinc-100 px-4 py-2 flex items-center justify-between">
+			<span class="font-semibold text-sm">📄 Week_4_Streamlit_RAGChatbot.pdf </span>
+			<div class="flex space-x-2">
+				<a href="/Time-Series-Team-Hub/pdf/Week_4___Streamlit_RAGChatbot.pdf" target="_blank" class="text-blue-600 hover:text-blue-800 text-sm">
+					Open in New Tab
+				</a>
+				<a href="/Time-Series-Team-Hub/pdf/Week_4___Streamlit_RAGChatbot.pdf" download class="text-blue-600 hover:text-blue-800 text-sm">
+					Download
+				</a>
+			</div>
+		</div>
+		<iframe 
+			src="/Time-Series-Team-Hub/pdf/Week_4___Streamlit_RAGChatbot.pdf" 
+			width="100%" 
+			height="700px"
+			class="border-0"
+			title="Week_4_Streamlit_RAGChatbot.pdf">
+			<p>Your browser does not support PDFs. <a href="/Time-Series-Team-Hub/pdf/Week_4___Streamlit_RAGChatbot.pdf">Download the PDF</a>.</p>
+		</iframe>
+	</div>
+</details>
+
+<details>
+	<summary> Week_4_ChatbotSlide.pdf </summary>
+	<div class="pdf-container my-6 border rounded-lg overflow-hidden shadow-lg">
+		<div class="pdf-header bg-zinc-100 px-4 py-2 flex items-center justify-between">
+			<span class="font-semibold text-sm">📄 Week_4_ChatbotSlide.pdf </span>
+			<div class="flex space-x-2">
+				<a href="/Time-Series-Team-Hub/pdf/Week_4___RAGChatbotSlide.pdf" target="_blank" class="text-blue-600 hover:text-blue-800 text-sm">
+					Open in New Tab
+				</a>
+				<a href="/Time-Series-Team-Hub/pdf/Week_4___RAGChatbotSlide.pdf" download class="text-blue-600 hover:text-blue-800 text-sm">
+					Download
+				</a>
+			</div>
+		</div>
+		<iframe 
+			src="/Time-Series-Team-Hub/pdf/Week_4___RAGChatbotSlide.pdf" 
+			width="100%" 
+			height="700px"
+			class="border-0"
+			title="Week_4_ChatbotSlide.pdf">
+			<p>Your browser does not support PDFs. <a href="/Time-Series-Team-Hub/pdf/Week_4___RAGChatbotSlide.pdf">Download the PDF</a>.</p>
+		</iframe>
+	</div>
+</details>
 </p>
-<hr>
+<h2 id="week_4___streamlit_ragchatbot">Week_4___Streamlit_RAGChatbot</h2>
 <p>🧠 <em>Repository managed by <a href="https://github.com/Jennifer1907/Time-Series-Team-Hub">Time Series Team Hub</a></em></p>
